requirements.txt
flask
click==6
webargs
marshmallow
pyyaml